Refactored day-of to use NextOfKinDisplay component with page wrapper.
