+++ src/applications/check-in/day-of/pages/NextOfKin.jsx
@@ -9,8 +9,8 @@ import BackButton from '../components/BackButton';
 import BackToHome from '../components/BackToHome';
 import { focusElement } from 'platform/utilities/ui';
 import Footer from '../components/Footer';
-import DemographicItem from '../../components/DemographicItem';
 import { seeStaffMessageUpdated } from '../actions';
+import NextOfKinDisplay from '../../components/pages/nextOfKin/NextOfKinDisplay';
 
 const NextOfKin = props => {
   const {
@@ -56,13 +56,6 @@ const NextOfKin = props => {
     [router, updateSeeStaffMessage],
   );
 
-  const nextOfKinFields = [
-    { title: 'Name', key: 'name' },
-    { title: 'Relationship', key: 'relationship' },
-    { title: 'Address', key: 'address' },
-    { title: 'Phone', key: 'phone' },
-    { title: 'Work phone', key: 'workPhone' },
-  ];
   if (isLoading) {
     return <LoadingIndicator message={'Loading your appointments for today'} />;
   } else if (!nextOfKin) {
@@ -70,49 +63,18 @@ const NextOfKin = props => {
     return <></>;
   } else {
     return (
-      <div className="vads-l-grid-container vads-u-padding-bottom--6 vads-u-padding-top--2 check-in-next-of-kin">
+      <>
         {(isUpdatePageEnabled || isDemographicsPageEnabled) && (
           <BackButton router={router} />
         )}
-        <h1>Is this your current next of kin information?</h1>
-        <div className="vads-u-border-color--primary vads-u-border-left--5px vads-u-margin-left--0p5 vads-u-padding-left--2">
-          <dl>
-            {nextOfKinFields.map(nextOfKinField => (
-              <React.Fragment key={nextOfKinField.key}>
-                <dt className="vads-u-font-size--h3 vads-u-font-family--serif">
-                  {nextOfKinField.title}
-                </dt>
-                <dd>
-                  {nextOfKinField.key in nextOfKin &&
-                  nextOfKin[nextOfKinField.key] ? (
-                    <DemographicItem
-                      demographic={nextOfKin[nextOfKinField.key]}
-                    />
-                  ) : (
-                    'Not available'
-                  )}
-                </dd>
-              </React.Fragment>
-            ))}
-          </dl>
-        </div>
-        <button
-          onClick={() => yesClick()}
-          className={'usa-button-secondary'}
-          data-testid="yes-button"
-        >
-          Yes
-        </button>
-        <button
-          onClick={() => noClick()}
-          className="usa-button-secondary vads-u-margin-top--2"
-          data-testid="no-button"
-        >
-          No
-        </button>
-        <Footer />
+        <NextOfKinDisplay
+          nextOfKin={nextOfKin}
+          yesAction={yesClick}
+          noAction={noClick}
+          Footer={Footer}
+        />
         <BackToHome />
-      </div>
+      </>
     );
   }
 };
